requirements.txt
streamlit
transformers
torch
pyspellchecker
